Add graphite web service and deployment
The graphite querier relies on a graphite web stock deployment to
resolve the query functions that are not yet supported by the querier.

For this reason a stock graphite deployment is required to run side by
side with the graphite querier when the graphite functionality is
enabled.

Also start working on the graphite querier